+++ conf/portal.application.conf
@@ -175,7 +175,7 @@ reporting {
     }
     GRAFANA {
       "text/html; charset=utf-8" {
-        type = "com.arpnetworking.metrics.portal.reports.impl.chrome.grafana.HtmlScreenshotRenderer"
+        type = "com.arpnetworking.metrics.portal.reports.impl.chrome.grafana.HtmlGrafanaScreenshotRenderer"
         chromePath = "/usr/lib/chromium/chrome"
         # ^ This path is only correct in the MPortal Docker image! If you aren't running MPortal in Docker, and you have reports enabled,
         #   you'll need to change this to the actual path to your Chrome binary.
@@ -189,7 +189,7 @@ reporting {
         }
       }
       "application/pdf" {
-        type = "com.arpnetworking.metrics.portal.reports.impl.chrome.grafana.PdfScreenshotRenderer"
+        type = "com.arpnetworking.metrics.portal.reports.impl.chrome.grafana.PdfGrafanaScreenshotRenderer"
         chromePath = "/usr/lib/chromium/chrome"
         # ^ This path is only correct in the MPortal Docker image! If you aren't running MPortal in Docker, and you have reports enabled,
         #   you'll need to change this to the actual path to your Chrome binary.
